Tarea Corta 1/TC1/docker/FlaskApp/app/requirements.txt
flask
psycopg2
pandas
mysql-connector-python
elasticsearch